include/types.h: move sha_digest_t to separate header
