+++ src/include/types.cc
@@ -14,6 +14,7 @@
  */
 
 #include "types.h"
+#include "common/Formatter.h"
 
 #include <iomanip>
