+++ src/msg/async/frames_v2.h
@@ -5,7 +5,7 @@
 #include "include/byteorder.h" // for ceph_le*
 #include "include/ceph_features.h" // for CEPH_FEATUREMASK_*
 #include "include/intarith.h" // for p2roundup()
-#include "include/types.h" // for sha256_digest_t
+#include "include/sha_digest.h" // for sha256_digest_t
 #include "common/Clock.h"
 #include "crypto_onwire.h"
 #include "compression_onwire.h"
